dialog: set the end time stamp immediately after going into deleted state
- checks to see if time stamps are set when testing if dialogs lasted
  too long in intermediary states

(cherry picked from commit 74825c896dacd3f39948395cebfcf7b6e0730b51)

+++ src/modules/dialog/dlg_hash.c
@@ -229,21 +229,24 @@ int dlg_clean_run(ticks_t ti)
 		while (dlg) {
 			tdlg = dlg;
 			dlg = dlg->next;
-			if(tdlg->state==DLG_STATE_UNCONFIRMED && tdlg->init_ts<tm-300) {
+			if(tdlg->state==DLG_STATE_UNCONFIRMED  && tdlg->init_ts>0
+					&& tdlg->init_ts<tm-300) {
 				/* dialog in early state older than 5min */
 				LM_NOTICE("dialog in early state is too old (%p ref %d)\n",
 						tdlg, tdlg->ref);
 				unlink_unsafe_dlg(&d_table->entries[i], tdlg);
 				destroy_dlg(tdlg);
 			}
-			if(tdlg->state==DLG_STATE_CONFIRMED_NA && tdlg->start_ts<tm-60) {
+			if(tdlg->state==DLG_STATE_CONFIRMED_NA && tdlg->start_ts>0
+					&& tdlg->start_ts<tm-60) {
 				if(update_dlg_timer(&tdlg->tl, 10)<0) {
 					LM_ERR("failed to update dialog lifetime in long non-ack state\n");
 				}
 				tdlg->lifetime = 10;
 				tdlg->dflags |= DLG_FLAG_CHANGED;
 			}
-			if(tdlg->state==DLG_STATE_DELETED && tdlg->end_ts<tm-300) {
+			if(tdlg->state==DLG_STATE_DELETED && tdlg->end_ts>0
+					&& tdlg->end_ts<tm-300) {
 				/* dialog in deleted state older than 5min */
 				LM_NOTICE("dialog in delete state is too old (%p ref %d)\n",
 						tdlg, tdlg->ref);
